// src/moduleA/lib/jl_lib/jl_ota_2.1.1.d.ts
declare enum UpgradeType {
    UPGRADE_TYPE_UNKNOWN = -1,
    UPGRADE_TYPE_CHECK_FILE = 0,
    UPGRADE_TYPE_FIRMWARE = 1
}
/** 文件偏移 */
declare class FileOffset {
    readonly offset: number;
    readonly len: number;
    constructor(offset: number | undefined, len: number | undefined);
    toString(): string;
}
/** OTA 配置 */
declare class OTAConfig {
    static readonly COMMUNICATION_WAY_BLE = 0;
    static readonly COMMUNICATION_WAY_SPP = 1;
    static readonly COMMUNICATION_WAY_USB = 2;
    /** 通讯方式*/
    communicationWay: number;
    /** 是否支持新的回连方式*/
    isSupportNewRebootWay: boolean;
    /** 固件升级文件数据*/
    updateFileData?: Uint8Array;
    toString(): string;
}
/** 回连信息*/
declare class ReConnectMsg {
    isSupportNewReconnectADV?: boolean;
    /** rcsp协议中的BLE_Mac地址  */
    deviceBleMac?: string;
    copy(): ReConnectMsg;
    toString(): string;
}
// ...
interface Logger {
    logv: (tag: string, ...args: any[]) => void;
    logd: (tag: string, ...args: any[]) => void;
    logi: (tag: string, ...args: any[]) => void;
    logw: (tag: string, ...args: any[]) => void;
    loge: (tag: string, ...args: any[]) => void;
}